third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/block-container-block-end-self-collapsing-children-offsets-nested-multiple-times-expected.txt

This is a testharness.js-based test.
[FAIL] .collapsed 1
  assert_equals: \n<item data-offset-y="58" class="collapsed"></item>\noffsetTop expected 58 but got 108
Harness: the test ran to completion.
